The great en-docs-ening
This documnts every existing public item, even if just with a single
sentence. There's room for improvement here, and especially room for
more details of how the different layers of the crate are intended to
work (e.g., how does a teensy_32::io::Serial wrap a mk20dx256::Uart)

Also removes mention of boards that won't be supported in the first
release, because for some reason I stopped to do that while I was
halfway through working on docs.