Fix Red-V QSPI divider when clock rate is decreased
Previously, the QSPI clock divider was always set before the PLL is
updated. This is correct when the clock rate is being increased, since
the new divider will just cause the QSPI to run slowly until the PLL
is updated.

When decreasing the clock speed, though, the new dividers could cause
the QSPI to be clocked faster than is stable, leading to
crashes. Instead, when the clock rate is being decreased, we should
update the QPI divisor after the PLL has been slowed.