Allow specifying options at serial enable time, with an Invert option
This allows the invert flag to be specified on boards that support
it (just the Teensies, for now).

This also DRYs up a bit of the serial init code, since I had to touch
a lot of places to do this. Future options will be much smaller patch
sets.

I also noticed one of the Teensy LC UART clock calculations was wrong,
and fixed that too.