Only allow getting handles to peripherals for current MCU
The guts of the change is to add a new trait to every non-gated
peripheral, which allows us to conditionally compile `get` functions
for everything.

Doing this meant renaming the existing `Peripheral` trait to
`GatedPeripheral` on kinetis. The `Peripheral` trait for SiFive
already served this purpose.

Also fixes a couple of missing `Drop` impls for sifive
peripherals. This needed some gnarly `ManuallyDrop` shenanigans for
the UART, unfortunately. I'm not totally happy with how that turned
out, but it works.

Fixes #1