Add MIDI to Control Voltage example
This demo takes a MIDI input and converts note on/off messages to
analog pitch and velocity signals using a couple of SPI DACs. With
another stage to scale these outputs, they could be used as inputs to
a VCO + VCA of a synthesizer. The MCP4921 DACs used in the demo
probably aren't totally ideal for this application, but they're what I
had on hand.