Add target-cpu/target-feature for suppported boards
`cargo cntrlr` will now set the target CPU and any target features for
the selected board/mcu. These can be overriden using the "RUSTFLAGS"
environment variable.

The "arduinouno" board still selects a non-existent target, but it's a
DIFFERENT non-existent target now. There's a plan for this, but with
AVR on hold I'm not gonna mess with it just yet.