mosi -> sdo, miso -> sdi
SDI/SDO appears to be the preferred current nomenclature, and better
reflects how the Kinetis SPI actually works - the out is always out
and the in is always in, regardless of if the SPI is configured as a
controller or peripheral.

SCK and CS were already fine. None of these will conflict with I2C lines,
which will be SCL and SDA.